requirements.txt
bigdl-llm[all]
python-pptx~=0.6.22
